Refactoring with module pattern
