+++ runner.rb
@@ -12,4 +12,4 @@ locations = {
 }
 
 stat_tracker = StatTracker.from_csv(locations)
-
+require 'pry'; binding.pry
